backend/requirements.txt
pydantic-ai==0.8.1
requests
python-dotenv
pydantic
